# openapi.yaml
openapi: 3.0.3
info:
  title: Mini CRM — API
  version: "1.0.0"
  description: |
    API do Mini CRM (consultoria / mentorias / IA).

    ## Autenticação (dois planos)
    - **API key (máquina/IA)** — envie `Authorization: Bearer <API_KEY>`. Escritas ficam
      registradas como `created_by = "ia"`. É o caminho recomendado para automações.
      Use o botão **Authorize** aqui no topo para testar com sua key.
    - **Sessão (navegador)** — `POST /api/auth/login` cria um cookie `crm_sessao` (httpOnly) e
      devolve um token **CSRF** que deve ir no header `X-CSRF-Token` em toda escrita. Escritas
      ficam como `created_by = "humano"`.

    As rotas de **Usuários** (`/api/usuarios`) e **API Keys** (`/api/keys`) exigem
    **admin via sessão** — uma Bearer key recebe `403`.

    ## Papéis
    - **admin** — administra usuários e chaves de API, além de usar o CRM.
    - **assistente** — usa o CRM inteiro (criar/editar/excluir leads e interações),
      mas não administra acessos.

    Todos os usuários enxergam a mesma carteira de leads. O papel é lido do banco a cada
    requisição, então promover ou rebaixar alguém vale na hora, sem precisar de novo login.

    ## Servidor MCP

    Além desta API REST, o CRM expõe um **servidor MCP** autenticado em `POST /mcp` (transporte
    Streamable HTTP) para agentes de IA, usando a **mesma chave Bearer**. REST e MCP compartilham a
    lógica de negócio em `crm-service.js` (comportamento idêntico). Veja `docs/MCP.md`.
servers:
...
security:
  - bearerAuth: []
paths:
  /api/auth/login:
    post:
      tags: [Autenticação]
      summary: Login (cria sessão)
      description: Define o cookie `crm_sessao` e devolve o token CSRF para as escritas via navegador.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: Autenticado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401':
          $ref: '#/components/responses/Erro401'
  /api/auth/logout:
    post:
      tags: [Autenticação]
      summary: Logout (destrói a sessão)
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '401': { $ref: '#/components/responses/Erro401' }
  /api/auth/me:
    get:
      tags: [Autenticação]
      summary: Sessão atual
      description: Requer cookie de sessão (não funciona com Bearer key).
      responses:
        '200':
          description: Dados da sessão
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
        '401': { $ref: '#/components/responses/Erro401' }
  /api/auth/senha:
    put:
      tags: [Autenticação]
      summary: Trocar a própria senha
      description: |
        Requer sessão (uma Bearer key recebe `403`). Exige a senha atual. Ao trocar, as **outras**
        sessões do usuário são encerradas; a sessão que fez a chamada continua válida.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TrocaSenhaRequest' }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/Erro400' }
        '401': { $ref: '#/components/responses/Erro401' }
        '403':
          description: Senha atual incorreta, ou credencial que não é de sessão
  /api/usuarios:
    get:
      tags: [Usuários]
      summary: Listar usuários
      description: Somente admin via sessão. Nunca devolve o hash da senha.
      responses:
        '200':
          description: Lista de usuários
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Usuario' }
        '401': { $ref: '#/components/responses/Erro401' }
        '403': { $ref: '#/components/responses/Erro403Admin' }
    post:
      tags: [Usuários]
      summary: Cadastrar usuário
      description: |
        Somente admin via sessão. O admin define a senha inicial (mínimo 8 caracteres) e a
        combina com a pessoa, que pode trocá-la depois em `PUT /api/auth/senha`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NovoUsuarioRequest' }
      responses:
        '201':
          description: Usuário criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Usuario' }
        '400': { $ref: '#/components/responses/Erro400' }
        '403': { $ref: '#/components/responses/Erro403Admin' }
        '409':
          description: Já existe um usuário com este e-mail
  /api/usuarios/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    put:
      tags: [Usuários]
      summary: Atualizar nome e/ou papel
      description: |
        Somente admin via sessão. O sistema recusa rebaixar o **último** administrador,
        para nunca ficar sem quem administre.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AtualizaUsuarioRequest' }
      responses:
        '200':
          description: Usuário atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Usuario' }
        '400': { $ref: '#/components/responses/Erro400' }
        '403': { $ref: '#/components/responses/Erro403Admin' }
        '404': { $ref: '#/components/responses/Erro404' }
    delete:
      tags: [Usuários]
      summary: Excluir usuário
      description: |
        Somente admin via sessão. As sessões do usuário caem junto. Recusa excluir a própria
        conta ou o último administrador. Os leads que ele cadastrou continuam no CRM; as chaves
        de API que ele criou seguem ativas (revogue-as separadamente).
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/Erro400' }
        '403': { $ref: '#/components/responses/Erro403Admin' }
        '404': { $ref: '#/components/responses/Erro404' }
  /api/usuarios/{id}/senha:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    put:
      tags: [Usuários]
      summary: Redefinir a senha de um usuário
      description: |
        Somente admin via sessão — para quando a pessoa perdeu o acesso. **Todas** as sessões
        dela são encerradas. Não exige a senha antiga (diferente de `PUT /api/auth/senha`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [senha]
              properties:
                senha: { type: string, minLength: 8, example: "senhaProvisoria2026" }
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '400': { $ref: '#/components/responses/Erro400' }
        '403': { $ref: '#/components/responses/Erro403Admin' }
        '404': { $ref: '#/components/responses/Erro404' }
  /api/clientes:
    get:
      tags: [Leads]
      summary: Listar leads
      description: Retorna todos os leads, mais recentes primeiro, com a contagem de interações.
      responses:
        '200':
          description: Lista de leads
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Lead' }
        '401': { $ref: '#/components/responses/Erro401' }
    post:
      tags: [Leads]
      summary: Cadastrar lead novo
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NovoLead' }
      responses:
        '201':
          description: Lead criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lead' }
        '400': { $ref: '#/components/responses/Erro400' }
        '401': { $ref: '#/components/responses/Erro401' }
        '403': { $ref: '#/components/responses/Erro403Csrf' }
  /api/clientes/{id}:
    parameters:
      - $ref: '#/components/parameters/IdLead'
    get:
      tags: [Leads]
      summary: Detalhe do lead (com interações)
      responses:
        '200':
          description: Lead com histórico de interações
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Lead'
                  - type: object
                    properties:
                      interacoes:
                        type: array
                        items: { $ref: '#/components/schemas/Interacao' }
        '401': { $ref: '#/components/responses/Erro401' }
        '404': { $ref: '#/components/responses/Erro404' }
    put:
      tags: [Leads]
      summary: Atualizar lead
      description: Campos omitidos mantêm o valor atual.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AtualizaLead' }
      responses:
        '200':
          description: Lead atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lead' }
        '401': { $ref: '#/components/responses/Erro401' }
        '403': { $ref: '#/components/responses/Erro403Csrf' }
        '404': { $ref: '#/components/responses/Erro404' }
    delete:
      tags: [Leads]
      summary: Excluir lead (LGPD)
      description: Remove o lead e suas interações (cascade).
      responses:
        '200':
          description: Removido
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean, example: true }
                  removido: { type: integer, example: 29 }
        '401': { $ref: '#/components/responses/Erro401' }
        '403': { $ref: '#/components/responses/Erro403Csrf' }
        '404': { $ref: '#/components/responses/Erro404' }
  /api/clientes/{id}/etapa:
    parameters:
      - $ref: '#/components/parameters/IdLead'
    put:
      tags: [Leads]
      summary: Mover etapa / definir resultado
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/MudaEtapa' }
      responses:
        '200':
          description: Lead atualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Lead' }
        '400': { $ref: '#/components/responses/Erro400' }
        '401': { $ref: '#/components/responses/Erro401' }
        '403': { $ref: '#/components/responses/Erro403Csrf' }
        '404': { $ref: '#/components/responses/Erro404' }
  /api/clientes/{id}/export:
    parameters:
      - $ref: '#/components/parameters/IdLead'
    get:
      tags: [Leads]
      summary: Exportar lead (LGPD)
      description: Devolve o lead com as interações como anexo JSON (`Content-Disposition`).
      responses:
        '200':
          description: JSON do lead + interações
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Lead'
                  - type: object
                    properties:
                      interacoes:
                        type: array
                        items: { $ref: '#/components/schemas/Interacao' }
        '401': { $ref: '#/components/responses/Erro401' }
        '404': { $ref: '#/components/responses/Erro404' }
  /api/clientes/{id}/interacoes:
    parameters:
      - $ref: '#/components/parameters/IdLead'
    get:
      tags: [Interações]
      summary: Listar interações do lead
      responses:
        '200':
          description: Lista de interações (mais recentes primeiro)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Interacao' }
        '401': { $ref: '#/components/responses/Erro401' }
    post:
      tags: [Interações]
      summary: Registrar interação
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NovaInteracao' }
      responses:
        '201':
          description: Interação criada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Interacao' }
        '400': { $ref: '#/components/responses/Erro400' }
        '401': { $ref: '#/components/responses/Erro401' }
        '403': { $ref: '#/components/responses/Erro403Csrf' }
        '404': { $ref: '#/components/responses/Erro404' }
  /api/clientes/{id}/resumos:
    parameters:
      - $ref: '#/components/parameters/IdLead'
    post:
      tags: [Resumo de reunião]
      operationId: criarRascunhoResumo
      summary: Extrair rascunho de uma transcrição
      description: >-
        Envia a transcrição a um serviço de IA e devolve um rascunho para revisão.
        **Não grava nada no histórico do cliente** — para gravar, confirme o rascunho.
        Esta rota aceita corpo de até 512 KB (as demais seguem em 64 KB) e tem limite de
        requisições próprio, porque cada chamada aciona um serviço externo pago.
        Telefones, e-mails e documentos são mascarados antes do envio ao modelo.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/NovoRascunhoResumo' }
            example:
              transcricao: "Maria: obrigada pela apresentação. Ficou definido que seguimos com o diagnóstico..."
      responses:
        '201':
          description: Rascunho criado, aguardando revisão
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RascunhoResumo' }
        '400': { $ref: '#/components/responses/Erro400' }
        '401': { $ref: '#/components/responses/Erro401' }
        '403': { $ref: '#/components/responses/Erro403Csrf' }
        '404': { $ref: '#/components/responses/Erro404' }
        '413': { $ref: '#/components/responses/Erro413' }
        '429': { $ref: '#/components/responses/Erro429' }
        '503': { $ref: '#/components/responses/Erro503Ia' }
  /api/resumos/{id}:
    parameters:
      - $ref: '#/components/parameters/IdRascunho'
    get:
      tags: [Resumo de reunião]
      operationId: obterRascunhoResumo
      summary: Ler o rascunho em revisão
      description: >-
        Só o dono do rascunho o enxerga — a posse é da credencial que o criou.
        Credencial diferente recebe `404`, nunca `403`: confirmar a existência de um
        rascunho alheio já seria vazamento.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Rascunho
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RascunhoResumo' }
        '401': { $ref: '#/components/responses/Erro401' }
        '404': { $ref: '#/components/responses/Erro404' }
    delete:
      tags: [Resumo de reunião]
      operationId: descartarRascunhoResumo
      summary: Descartar o rascunho
      description: Irreversível. Nada é gravado no histórico e a transcrição em revisão é apagada.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '204': { description: Descartado }
        '401': { $ref: '#/components/responses/Erro401' }
        '403': { $ref: '#/components/responses/Erro403Csrf' }
        '404': { $ref: '#/components/responses/Erro404' }
  /api/resumos/{id}/confirmar:
    parameters:
      - $ref: '#/components/parameters/IdRascunho'
    post:
      tags: [Resumo de reunião]
      operationId: confirmarRascunhoResumo
      summary: Confirmar o rascunho revisado
      description: >-
        Grava no histórico o conteúdo **como revisado** — não a versão que o modelo produziu.
        A autoria e a marca de revisão vêm da credencial: sessão grava `revisao: humana`,
        chave de API grava `revisao: sem_revisao`. Nenhum campo de negócio do cliente é
        alterado, exceto a próxima ação quando `promover_proxima_acao` for enviado
        explicitamente.
      security:
        - cookieAuth: []
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfirmacaoResumo' }
            example:
              resumo: "Maria aprovou o diagnóstico para o time de atendimento."
              decisoes: [{ texto: "Seguir com o diagnóstico" }]
              proximos_passos: [{ texto: "Enviar proposta", responsavel: "Farley", prazo: "2026-09-05" }]
              objecoes: [{ texto: "Achou o prazo apertado" }]
              promover_proxima_acao: { texto: "Enviar proposta", data: "2026-09-05" }
      responses:
        '201':
          description: Registro criado no histórico
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Interacao' }
        '400': { $ref: '#/components/responses/Erro400' }
        '401': { $ref: '#/components/responses/Erro401' }
        '403': { $ref: '#/components/responses/Erro403Csrf' }
        '404': { $ref: '#/components/responses/Erro404' }
        '409': { $ref: '#/components/responses/Erro409ProximaAcao' }
  /api/interacoes/{id}/transcricao:
    parameters:
      - $ref: '#/components/parameters/IdInteracao'
    get:
      tags: [Resumo de reunião]
      operationId: obterTranscricao
      summary: Ler a transcrição de origem
      description: >-
        A transcrição é retida por 90 dias a partir da gravação e depois descartada
        automaticamente. Passado o prazo a resposta é `200` com `disponivel: false` —
        retenção cumprida não é erro. A transcrição **não** é incluída na exportação de
        dados do titular.
      security:
        - cookieAuth: []
        - bearerAuth: []
      responses:
        '200':
          description: Transcrição disponível, ou já descartada pela retenção
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TranscricaoResposta' }
        '401': { $ref: '#/components/responses/Erro401' }
        '404': { $ref: '#/components/responses/Erro404' }
  /api/hoje:
    get:
      tags: [Agenda]
      summary: Agenda "Hoje" (follow-ups)
      description: Leads em aberto com `proxima_acao_data`, agrupados em atrasados / hoje / futuros.
      responses:
        '200':
          description: Follow-ups agrupados
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HojeResponse' }
        '401': { $ref: '#/components/responses/Erro401' }
...
  /api/keys/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: integer }
    delete:
      tags: [API Keys]
      summary: Revogar API key
      description: Somente admin via sessão. Desativa a key (`ativa = 0`).
      responses:
        '200': { $ref: '#/components/responses/Ok' }
        '403': { $ref: '#/components/responses/Erro403Admin' }
        '404': { $ref: '#/components/responses/Erro404' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "API key da integração (formato: crm_xxxxxxxx...). Gerada por um admin."
    cookieAuth:
      type: apiKey
      in: cookie
      name: crm_sessao
      description: Cookie de sessão criado no login (uso pelo navegador; escritas exigem header X-CSRF-Token).
...
  responses:
...
    Erro401:
      description: Não autenticado / credencial inválida
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Erro' }
...